requirements.txt
fastapi
uvicorn
aiohttp
aiomysql
python-dotenv
openai
jinja2
